temp_clone/requirements.txt
pymongo==4.8.0
python-dateutil==2.9.0
rapidfuzz==3.9.6
python-dotenv==1.0.1